pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-template"
version = "0.1.0"
description = "Minimal generative AI template (DSPy + FastAPI)"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "dspy-ai",
    "openai",
    "fal-client",
]

[project.optional-dependencies]
server = ["fastapi", "uvicorn", "python-dotenv", "python-multipart", "google-cloud-storage"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = ["pytest"]
